crypto/cookiecutter/requirements.txt
pycryptodome>=3.18.0
flask==3.0.2